DOCS: update docs to match new architecture
